+++ py_trees_ros/conversions.py
@@ -26,8 +26,18 @@ import py_trees_msgs.msg as py_trees_msgs
 ##############################################################################
 
 
-def convert_type(behaviour):
-    # problems with decorators?
+def behaviour_type_to_msg_constant(behaviour):
+    """
+    Convert a behaviour class type to a message constant.
+
+    Args:
+        behaviour (:class:`~py_trees.behaviour.Behaviour`): investigate the type of this behaviour
+
+    Returns:
+        :obj:`uint8`: from the type constants in :class:`~py_trees_msgs.msg.Behaviour`
+    """
+    # TODO: problems with decorators?
+    # TODO: throw an exception if parent is not Behaviour?
     if isinstance(behaviour, py_trees.composites.Sequence):
         return py_trees_msgs.Behaviour.SEQUENCE
     elif isinstance(behaviour, py_trees.composites.Chooser):
@@ -39,10 +49,19 @@ def convert_type(behaviour):
     elif isinstance(behaviour, py_trees.behaviour.Behaviour):
         return py_trees_msgs.Behaviour.BEHAVIOUR
     else:
-        return 0  # unknown type
+        return py_trees_msgs.Behaviour.UNKNOWN_TYPE
 
 
-def convert_status(status):
+def status_enum_to_msg_constant(status):
+    """
+    Convert a status to a message constant.
+
+    Args:
+        status (:class:`~py_trees.common.Status`): status enum of a behaviour
+
+    Returns:
+        :obj:`uint8`: from the status constants in :class:`~py_trees_msgs.msg.Behaviour`
+    """
     if status == py_trees.common.Status.INVALID:
         return py_trees_msgs.Behaviour.INVALID
     elif status == py_trees.common.Status.RUNNING:
@@ -55,7 +74,16 @@ def convert_status(status):
         return 0  # unknown status
 
 
-def convert_blackbox_level(blackbox_level):
+def blackbox_enum_to_msg_constant(blackbox_level):
+    """
+    Convert a blackbox level enum to a message constant.
+
+    Args:
+        blackbox_level (:class:`~py_trees.common.BlackboxLevel`): blackbox level of a behaviour
+
+    Returns:
+        :obj:`uint8`: from the type constants in :class:`~py_trees_msgs.msg.Behaviour`
+    """
     if blackbox_level == py_trees.common.BlackBoxLevel.DETAIL:
         return py_trees_msgs.Behaviour.BLACKBOX_LEVEL_DETAIL
     elif blackbox_level == py_trees.common.BlackBoxLevel.COMPONENT:
@@ -67,22 +95,31 @@ def convert_blackbox_level(blackbox_level):
 
 
 def behaviour_to_msg(behaviour):
-    new_behaviour = py_trees_msgs.Behaviour()
-    new_behaviour.name = behaviour.name
-    new_behaviour.class_name = str(behaviour.__module__) + '.' + str(type(behaviour).__name__)
-    new_behaviour.own_id = unique_id.toMsg(behaviour.id)
-    new_behaviour.parent_id = unique_id.toMsg(behaviour.parent.id) if behaviour.parent else uuid_msgs.UniqueID()
-    new_behaviour.child_ids = [unique_id.toMsg(child.id) for child in behaviour.iterate(direct_descendants=True) if not child.id == behaviour.id]
+    """
+    Convert a behaviour to a message.
+
+    Args:
+        behaviour (:class:`~py_trees.behaviour.Behaviour`): behaviour to convert
+
+    Returns:
+        :class:`~py_trees_msgs.msg.Behaviour`: a ros message representation of a behaviour
+    """
+    msg = py_trees_msgs.Behaviour()
+    msg.name = behaviour.name
+    msg.class_name = str(behaviour.__module__) + '.' + str(type(behaviour).__name__)
+    msg.own_id = unique_id.toMsg(behaviour.id)
+    msg.parent_id = unique_id.toMsg(behaviour.parent.id) if behaviour.parent else uuid_msgs.UniqueID()
+    msg.child_ids = [unique_id.toMsg(child.id) for child in behaviour.iterate(direct_descendants=True) if not child.id == behaviour.id]
 
     tip = behaviour.tip()
     # tip_id is empty if the behaviour is invalid or if it is a valid
     # leaf
     if tip is not None and tip != behaviour:
-        new_behaviour.tip_id = unique_id.toMsg(tip.id)
+        msg.tip_id = unique_id.toMsg(tip.id)
 
-    new_behaviour.type = convert_type(behaviour)
-    new_behaviour.blackbox_level = convert_blackbox_level(behaviour.blackbox_level)
-    new_behaviour.status = convert_status(behaviour.status)
-    new_behaviour.message = behaviour.feedback_message
+    msg.type = behaviour_type_to_msg_constant(behaviour)
+    msg.blackbox_level = blackbox_enum_to_msg_constant(behaviour.blackbox_level)
+    msg.status = status_enum_to_msg_constant(behaviour.status)
+    msg.message = behaviour.feedback_message
 
-    return new_behaviour
+    return msg